+++ site/dat/docs/Changelog.md
@@ -6,7 +6,8 @@
  - add `delete-test-files` option to cloud provisioning
  - fix reading piped config from stdin
  - don't trap KeyboardInterrupt in tool install
- - add compressed folder sending support for remote and cloud provisioning  
+ - add [logic blocks](JMeter.md#Logic-Blocks) to `scenario.requests` syntax for JMeter
+ - remove default xmx set for JMeter
 
 ## 1.5.0 <sup>4 may 2016</sup>
  - add [Tsung](Tsung.md) executor